RecipeMaster/requirements.txt
Flask
Flask-SQLAlchemy
Werkzeug
flask-jwt-extended
flask-mail
python-dotenv
flask-wtf